+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 - Collection / Albums / Countries / Tags: added label when dragging scrollbar thumb
 - Albums: localized common album names
 - Collection: select shortcut icon image
+- Viewer: customizable quick actions
 
 ### Changed
 - Upgraded Flutter to beta v2.1.0-12.2.pre